rag-python/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
aiokafka==0.11.0
elasticsearch==8.16.0
minio==7.2.10
httpx==0.27.2
pydantic==2.10.2
pypdf==5.1.0
python-dotenv==1.0.1
